requirements.txt
pillow==10.2.0
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
matplotlib==3.8.2
seaborn==0.13.1
tqdm==4.66.1
setuptools>=68.2.2
wheel>=0.41.2
